April-30th HomeTab and BookingTab design - HomeTab: Moving HorizontalScrollView to BookingTab, adding intro/description/info - BookingTab: Implement HorizontalScrollView
