+++ app/src/main/res/layout/fragment_booking.xml
@@ -4,12 +4,103 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fillViewport="true"
+    android:orientation="vertical"
     tools:context=".appUI.BookingFragment">
 
-    <TextView
+    <LinearLayout
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:gravity="center"
-        android:text="Booking" />
+        android:layout_height="wrap_content"
+        android:orientation="vertical">
+
+        <LinearLayout
+            android:layout_marginTop="10dp"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:orientation="vertical"
+            android:layout_gravity="center">
+            <ImageView
+                android:layout_width="250dp"
+                android:layout_height="50dp"
+                android:background="@drawable/mtb_header"/>
+        </LinearLayout>
+
+        <TextView
+            android:layout_marginTop="10dp"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:padding="5dp"
+            android:text="Featured Movies"
+            android:textColor="#A06919"
+            android:textSize="20sp"
+            android:textStyle="bold" />
+        <View
+            android:layout_marginTop="5dp"
+            android:layout_width="match_parent"
+            android:layout_height="1dp"
+            android:background="#777777" />
+
+        <HorizontalScrollView
+            android:id="@+id/horizontalScrollView"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:scrollbars="none"
+            android:requiresFadingEdge="horizontal"
+            android:fadingEdgeLength="30dp">
+
+            <LinearLayout
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:padding="5dp"
+                android:orientation="horizontal">
+
+                <ImageView
+                    android:id="@+id/banner1"
+                    android:clickable="true"
+                    android:focusable="true"
+                    android:layout_width="110dp"
+                    android:layout_height="170dp"
+                    android:layout_margin="8dp"
+                    android:src="@drawable/poster_86" />
+
+                <ImageView
+                    android:id="@+id/banner2"
+                    android:clickable="true"
+                    android:focusable="true"
+                    android:layout_width="120dp"
+                    android:layout_height="170dp"
+                    android:layout_margin="8dp"
+                    android:src="@drawable/poster_josee" />
+
+                <ImageView
+                    android:id="@+id/banner3"
+                    android:clickable="true"
+                    android:focusable="true"
+                    android:layout_width="120dp"
+                    android:layout_height="170dp"
+                    android:layout_margin="8dp"
+                    android:src="@drawable/poster_violet" />
+
+                <ImageView
+                    android:id="@+id/banner4"
+                    android:clickable="true"
+                    android:focusable="true"
+                    android:layout_width="120dp"
+                    android:layout_height="170dp"
+                    android:layout_margin="8dp"
+                    android:src="@drawable/poster_yourname" />
+
+                <ImageView
+                    android:id="@+id/banner5"
+                    android:clickable="true"
+                    android:focusable="true"
+                    android:layout_width="120dp"
+                    android:layout_height="170dp"
+                    android:layout_margin="8dp"
+                    android:src="@drawable/poster_yourlie" />
+            </LinearLayout>
+        </HorizontalScrollView>
+
+    </LinearLayout>
+
 
 </ScrollView>
