+++ app/src/main/res/values/strings.xml
@@ -1,6 +1,10 @@
 <resources>
     <string name="app_name">Movie Tickets Booking</string>
     <string name="app_version">Version 1.0.0</string>
-    <!-- TODO: Remove or change this placeholder text -->
-    <string name="hello_blank_fragment">Hello blank fragment</string>
+    <string name="project_introduction">
+        Within a smartphone connecting to the Internet, we can pick a favorite movie,
+        select a suitable airing time and seats for a movie ticket booking at home.
+        This is a real time system connected to a clouding database, you can
+        know the status of the actual theater slots and make your own choice.
+    </string>
 </resources>
